chore(aws_sqs sink): migration to stream sink (#11708)
* migrate to stream sink


* split request builder from service


* fix tests
